Updates from QA thread
Grace Souter  1 hour ago
This is maybe picky, but it jumped out at me that the copy for Corolla refers to it as "there," followed up by the Navarre Beach copy that starts off with "Here". I think just swapping out "there" for "here" in the Corolla copy tightens it up.
I also wonder if we could swap "keep" for "limit" in the first paragraph of copy under the hero.
:this-darkmode:
1

Katelin Hill  23 minutes ago
Agreed with Grace on edits above
I think we should add 2022 in the intro after Best Places to Buy a Beach House so ppl know it's timely
Should the image in the bottom CTA be linked?
The rental revenue language is a little all over the board now. We changed "median annual gross rental revenue" everywhere but Corolla--to keep consistent and not be such a mouthful, can we say: ...The fact that a vacation rental property in Corolla could net you approximately $77,000 in rental revenue per year only shores up the investment.
Let's add "rental" before revenue in Gulf Shores: ...but can yield over $50,000 of rental revenue per year.
Let's add "in rental revenue...per" to New Buffalo: ...could earn you just over $58,000 in rental revenue per year.